# dolater-ios/DoLater/OpenAPI/openapi.yaml
openapi: "3.0.3"
info:
  title: Do Later API
  description: Do Later API
  version: 1.0.0
servers:
  - url: https://api.dolater.kantacky.com
    description: Production Server
  - url: https://stg.api.dolater.kantacky.com
    description: Staging Server
  - url: https://dev.api.dolater.kantacky.com
    description: Development Server
  - url: http://localhost:8080
    description: Local Server
paths:
...
  /users/{uid}:
    get:
      tags:
        - user
      summary: Get User
      description: Get a user
      operationId: getUser
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/uid"
      responses:
        200:
          $ref: "#/components/responses/UserOK"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags:
        - user
      summary: Update User
      description: Update a user
      operationId: updateUser
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/uid"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        200:
          $ref: "#/components/responses/UserOK"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - user
      summary: Delete User
      description: Delete a user
      operationId: deleteUser
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/uid"
      responses:
        204:
          $ref: "#/components/responses/NoContent"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
...
  /users/{uid}/follow:
    post:
      tags:
        - user
      summary: Follow User
      description: Follow a user
      operationId: followUser
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/uid"
      responses:
        200:
          $ref: "#/components/responses/FollowStatusOK"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - user
      summary: Unfollow User
      description: Unfollow a user
      operationId: unfollowUser
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/uid"
      responses:
        204:
          $ref: "#/components/responses/NoContent"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
...
  /tasks/{id}:
    get:
      tags:
        - task
      summary: Get Task
      description: Get a task
      operationId: getTask
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - task
      summary: Update Task Forcibly
      description: Update a task forcibly
      operationId: updateTaskForcibly
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        200:
          $ref: "#/components/responses/TaskOK"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    patch:
      tags:
        - task
      summary: Update Task
      description: Update a task
      operationId: updateTask
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        200:
          $ref: "#/components/responses/TaskOK"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - task
      summary: Delete Task
      description: Delete a task
      operationId: deleteTask
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        204:
          description: No Content
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
  /tasks/{id}/notify:
    post:
      tags:
        - task
      summary: Notify Task
      description: Notify a task
      operationId: notifyTask
      security:
        - AppCheck: []
        - AuthBearer: []
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotifyTaskInput"
      responses:
        204:
          $ref: "#/components/responses/NoContent"
        401:
          $ref: "#/components/responses/Unauthorized"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        photoURL:
          type: string
          format: uri
      required:
        - id
        - displayName
        - photoURL
    UpdateUserInput:
      type: object
      properties:
        displayName:
          type: string
        photoURL:
          type: string
          format: uri
      required: []
    FollowStatus:
      type: object
      properties:
        from:
          $ref: "#/components/schemas/User"
        to:
          $ref: "#/components/schemas/User"
        timestamp:
          type: string
          format: date-time
      required:
        - from
        - to
        - timestamp
    TaskPool:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner:
          $ref: "#/components/schemas/User"
        type:
          type: string
          enum:
            - taskPoolTypeActive
            - taskPoolTypeArchived
            - taskPoolTypeBin
            - taskPoolTypePending
      required:
        - id
        - type
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        removedAt:
          type: string
          format: date-time
        archivedAt:
          type: string
          format: date-time
        owner:
          $ref: "#/components/schemas/User"
        pool:
          $ref: "#/components/schemas/TaskPool"
      required:
        - id
        - url
        - createdAt
        - owner
        - pool
...
    UpdateTaskInput:
      type: object
      properties:
        url:
          type: string
          format: uri
        completedAt:
          type: string
          format: date-time
        removedAt:
          type: string
          format: date-time
        archivedAt:
          type: string
          format: date-time
        poolId:
          type: string
          format: uuid
      required: []
...
    NotifyTaskInput:
      type: object
      properties:
        emoji:
          type: string
      required:
        - emoji
  parameters:
    id:
      name: id
      in: path
      required: true
      description: ID
      schema:
        type: string
        format: uuid
    uid:
      name: uid
      in: path
      required: true
      description: UID
      schema:
        type: string
...
  securitySchemes:
    AuthBearer:
      type: http
      scheme: bearer
      description: Bearer Access Token for Authorization
    AppCheck:
      type: apiKey
      in: header
      name: X-Firebase-AppCheck
      description: Firebase App Check Token
...
  responses:
    UserOK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/User"
...
    FollowStatusOK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FollowStatus"
...
    TaskOK:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Task"
    NoContent:
      description: No Content
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Unauthorized
    NotFound:
      description: NotFound
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: NotFound
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: Internal Server Error